requirements.txt
aniso8601==9.0.1
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Caching==2.0.1
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlitedict==2.0.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
